Added in further Debugging Methods
Added in a Path In Progress Event to allow debugging of paths in real
time. There is a performance hit when active due to requiring to run in
a separate Task/Thread to allow rendering by MonoGame. So when doing
Perf Testing comment out the Task Run and unsubscribe event